requirements.txt
ccxt
python-dotenv
numpy
requests
gspread
oauth2client